requirements.txt
altair==4.2.0
attrs==22.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
ga-tools==0.2
gitdb==4.0.10
GitPython==3.1.29
graphviz==0.20.1
idna==3.4
importlib-metadata==5.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
numpy==1.24.1
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pipreqs==0.4.11
protobuf==4.21.12
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pyserial==3.5
python-dateutil==2.8.2
python-util==1.2.1
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.16.0
streamlit-ace==0.1.1
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
yarg==0.1.9
zipp==3.11.0

deta~=1.1.0
python-dotenv~=0.21.0
